Add files via upload
Fixed parenthesis handling, now having issues with expressions that aren't "correct"

For example the expression "200 *(_(1)*_(_5))", outputs the following:

STACK IS EMPTY
STACK IS EMPTY
1000

Need to figure out logic for handling that expression, or where my logic is incorrect. I also managed to even process a string like this: " 200*___()(((_5)))",  but it outputted 1000 instead of -1000.

Could someone please assist me with figuring out the logic to handle those strings?

Thank you!